cryptobot_production/requirements.txt
# Core Dependencies
ccxt==4.1.52
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0

# AI Components
openai==0.27.8
langchain==0.0.346

# Monitoring
prometheus-client==0.17.1
psutil==5.9.5

# Utilities
pyyaml==6.0
python-telegram-bot==20.3
aiohttp==3.8.5